refactor: simplify SoccerNet frame selection pipeline
